requirements/base.txt
Django==5.1
paho-mqtt==2.1.0